pyproject.toml
[tool.poetry]
name = "dlt-axie-infinity-demo"
version = "0.1.0"
description = "DLT Pipeline Script to get Axie Infinity smart contracts data on Ronin Network into BigQuery/Redshift"
authors = ["Marcin Rudolf <rudolfix@scalevector.ai>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8,<3.11"
web3 = "^5.30.0"
python-dlt = {git = "https://github.com/scale-vector/dlt.git", rev = "rfix/wei-type-variant-support", extras = ["gcp", "redshift"]}


[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
mypy = "0.971"
flake8 = "^5.0.0"
bandit = "^1.7.0"
flake8-bugbear = "^22.0.0"
types-simplejson = "^3.17.0"
types-requests = "^2.25.6"
types-python-dateutil = "^2.8.15"
pytest-pythonpath = "^0.7.4"
types-PyYAML = "^6.0.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
